Make compressed sizes 64 bit
Reading lots of clear codes in the Shrink decoder could possibly
lead to more than 4 GB consumed. Perhaps any such element should
be rejected, as no known Zip program supports both Zip64 and the
generation of legacy formats.

The change is done for Reduce and Implode as well, to cover all
our bases.

+++ libarchive/archive_read_support_format_zip.c
@@ -1740,7 +1740,7 @@ zip_read_data_implode(struct archive_read *a, const void **buff,
 {
 	struct zip *zip = (struct zip *)(a->format->data);
 	int r;
-	size_t cmp_size;
+	uint64_t cmp_size;
 
 	(void)offset; /* UNUSED */
 
@@ -1782,7 +1782,7 @@ zip_read_data_shrink(struct archive_read *a, const void **buff,
 {
 	struct zip *zip = (struct zip *)(a->format->data);
 	int r;
-	size_t cmp_size;
+	uint64_t cmp_size;
 
 	(void)offset; /* UNUSED */
 
@@ -1823,7 +1823,7 @@ zip_read_data_reduce(struct archive_read *a, const void **buff,
 {
 	struct zip *zip = (struct zip *)(a->format->data);
 	int r;
-	size_t cmp_size;
+	uint64_t cmp_size;
 
 	(void)offset; /* UNUSED */
